+ repo/psql -> исправлен тип для колонки delta (INTEGER -> BIGINT)

+++ internal/usecase/repo/server_postgre.go
@@ -18,7 +18,7 @@ const (
 CREATE TABLE IF NOT EXISTS metrics(
     "id" VARCHAR(255) UNIQUE NOT NULL,
     "type" VARCHAR(50) NOT NULL,
-    "delta" INTEGER,
+    "delta" BIGINT,
     "value" DOUBLE PRECISION,
     "hash" VARCHAR(255)
     );
